processes/version_2.1/requirements.txt
BeautifulSoup==3.2.1
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
chardet==3.0.4
click==6.7
coloredlogs==7.3
Flask==0.12.2
humanfriendly==4.4.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
lxml==4.1.1
MarkupSafe==1.0
monotonic==1.4
nltk==3.2.5
numpy==1.13.3
pika==0.11.2
pymongo==3.5.1
python-crontab==2.2.8
python-dateutil==2.6.1
redis==2.10.6
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
urllib3==1.22
Werkzeug==0.12.2
